refactor(model): deduplicate functions
