+++ mkdocs.yml
@@ -43,19 +43,19 @@ repo_url: https://github.com/PorametPat/inspeqtor
 plugins:
     - search
     - mkdocstrings:
-        handlers:
-            python:
-                options:
-                    members_order: source
-                    show_bases: false
-                    show_root_heading: true
-                    show_signature_annotations: true
-                    show_symbol_type_heading: true
-                    show_symbol_type_toc: true
-                    parameter_headings: false
-                    show_signature: true
-                    separate_signature: true
-                    signature_crossrefs: true
+          handlers:
+              python:
+                  options:
+                      members_order: source
+                      show_bases: false
+                      show_root_heading: true
+                      show_signature_annotations: true
+                      show_symbol_type_heading: true
+                      show_symbol_type_toc: true
+                      parameter_headings: false
+                      show_signature: true
+                      separate_signature: true
+                      signature_crossrefs: true
     - mkdocs-jupyter:
           remove_tag_config:
               remove_input_tags:
@@ -90,6 +90,7 @@ nav:
           - Generate dataset: tutorials/tutorial_0001_dataset.ipynb
           - C&C: tutorials/tutorial_0003_cc.ipynb
           - Control Action: tutorials/tutorial_0002_control.ipynb
+          - Marimo: tutorials/tutortial_0004_marimo.md
     - explanation.md
     - How-To Guides:
           - Generate dataset: guides/generate-dataset.md
