+++ src/inspeqtor/experimental/model.py
@@ -30,6 +30,7 @@ from .constant import (
     X,
     Y,
     Z,
+    default_expectation_values_order,
 )
 from .control import ControlSequence
 from .ctyping import Wos
@@ -386,7 +387,7 @@ def load_model(path: pathlib.Path | str, skip_history: bool = False):
     )
 
 
-def Wo_2_level_v3(U: jnp.ndarray, D: jnp.ndarray) -> jnp.ndarray:
+def hermitian(U: jnp.ndarray, D: jnp.ndarray) -> jnp.ndarray:
     """This is a function that parametrized Hermitian matrix
 
     Args:
@@ -461,7 +462,9 @@ def unitary(params: jnp.ndarray) -> jnp.ndarray:
     return psi_ * Q
 
 
-def get_spam(params: VariableDict):
+def get_spam(
+    params: VariableDict,
+) -> tuple[list[ExpectationValue], dict[str, jnp.ndarray]]:
     pair_map = {"+": "-", "-": "+", "0": "1", "1": "0", "r": "l", "l": "r"}
     observables = {"X": X, "Y": Y, "Z": Z}
     for pauli, matrix in observables.items():
@@ -494,6 +497,126 @@ def get_spam(params: VariableDict):
         expvals.append(expval)
 
     return expvals, observables
+
+
+def unitary_to_expvals(output, unitaries: jnp.ndarray) -> jnp.ndarray:
+    """Function to be used with Unitary-based model for probabilistic model construction
+    with `make_probabilistic_model` function.
+
+    Args:
+        output (typing.Any): The output from Unitary-based model
+        unitaries (jnp.ndarray): Ideal unitary, ignore in this function.
+
+    Returns:
+        jnp.ndarray: Expectation values array
+    """
+    U = unitary(output)
+    return get_predict_expectation_value(
+        {"X": X, "Y": Y, "Z": Z},
+        U,
+        default_expectation_values_order,
+    )
+
+
+def toggling_unitary_to_expvals(
+    output: jnp.ndarray, unitaries: jnp.ndarray
+) -> jnp.ndarray:
+    """Calculate $U_J$ and convert it to expectation values.
+
+    Args:
+        output (jnp.ndarray): Parameters parametrized unitary operator
+        unitaries (jnp.ndarray): Ideal unitary operators corresponding to the output
+
+    Returns:
+        jnp.ndarray: Predicted expectation values
+    """
+    UJ: jnp.ndarray = unitary(output)  # type: ignore
+    UJ_dagger = jnp.swapaxes(UJ, -2, -1).conj()
+
+    X_ = UJ_dagger @ X @ UJ
+    Y_ = UJ_dagger @ Y @ UJ
+    Z_ = UJ_dagger @ Z @ UJ
+
+    return get_predict_expectation_value(
+        {"X": X_, "Y": Y_, "Z": Z_},
+        unitaries,
+        default_expectation_values_order,
+    )
+
+
+def toggling_unitary_with_spam_to_expvals(
+    output, unitaries: jnp.ndarray
+) -> jnp.ndarray:
+    """To model the SPAM noise with probabilistic model and UJ model
+
+    Note:
+        Expected the output structure as follows
+        ```python
+        spam_params = {
+            "SP": {
+                "+": jnp.array([0.9]),
+                "-": jnp.array([0.9]),
+                "r": jnp.array([0.9]),
+                "l": jnp.array([0.9]),
+                "0": jnp.array([0.9]),
+                "1": jnp.array([0.9]),
+            },
+            "AM": {
+                "X": {"prob_10": jnp.array([0.1]), "prob_01": jnp.array([0.1])},
+                "Y": {"prob_10": jnp.array([0.1]), "prob_01": jnp.array([0.1])},
+                "Z": {"prob_10": jnp.array([0.1]), "prob_01": jnp.array([0.1])},
+            },
+        }
+
+        output = {
+            "model_params": ...,
+            "spam_params": spam_params
+        }
+        ```
+
+    Args:
+        output (typing.Any): _description_
+        unitaries (jnp.ndarray): _description_
+
+    Returns:
+        jnp.ndarray: _description_
+    """
+
+    model_output = output["model_params"]
+    spam_output = output["spam_params"]
+
+    UJ: jnp.ndarray = unitary(model_output)  # type: ignore
+    UJ_dagger = jnp.swapaxes(UJ, -2, -1).conj()
+
+    expectation_value_order, observables = get_spam(spam_output)
+
+    X_ = UJ_dagger @ observables["X"] @ UJ
+    Y_ = UJ_dagger @ observables["Y"] @ UJ
+    Z_ = UJ_dagger @ observables["Z"] @ UJ
+
+    return get_predict_expectation_value(
+        {"X": X_, "Y": Y_, "Z": Z_},
+        unitaries,
+        expectation_value_order,
+    )
+
+
+def observable_to_expvals(output, unitaries: jnp.ndarray) -> jnp.ndarray:
+    """Function to be used with Wo-based model for probabilistic model construction
+    with `make_probabilistic_model` function.
+
+    Args:
+        output (typing.Any): The output from Wo-based model
+        unitaries (jnp.ndarray): Ideal unitary, ignore in this function.
+
+    Returns:
+        jnp.ndarray: Expectation values array
+    """
+    return get_predict_expectation_value(
+        observable=output,
+        unitaries=unitaries,
+        evaluate_expectation_values=default_expectation_values_order,
+    )
 
 
 def model_parse_fn(key, value):
