+++ src/inspeqtor/experimental/models/__init__.py
@@ -1,1 +1,11 @@
 from . import nnx as nnx, linen as linen
+from ..model import (
+    mse as mse,
+    get_predict_expectation_value as get_predict_expectation_value,
+    calculate_metric as calculate_metric,
+    hermitian as hermitian,
+    unitary as unitary,
+    get_spam as get_spam,
+    ModelData as ModelData,
+    LossMetric as LossMetric,
+)
