--- src/inspeqtor/legacy/__init__.py
@@ -1,16 +1,0 @@
-from . import (
-    _pennylane as _pennylane,
-    data as data,
-    constant as constant,
-    pulse as pulse,
-    typing as typing,
-    utils as utils,
-    qiskit as qiskit,
-    physics as physics,
-    model as model,
-    optimizer as optimizer,
-    visualization as visualization,
-    decorator as decorator,
-    model_v2 as model_v2,
-    torch as torch,
-)
