--- src/inspeqtor/legacy/utils/__init__.py
@@ -1,1 +1,0 @@
-from . import predefined as predefined, helper as helper
